+++ CHANGELOG.md
@@ -13,6 +13,8 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Changed
 
+- The Gateway Server worker pools may now drop workers if they are idle for too long.
+
 ### Deprecated
 
 ### Removed